working

+++ common/hex_dump.cc
@@ -1,12 +1,12 @@
-
 #include "common/hex_dump.hh"
 #include "std/stdio.hh"
+#include "std/ctype.hh"
 
 namespace vt::common {
 
 void hex_dump(const void* ptr, size_t buflen) {
   auto buf = static_cast<const unsigned char*>(ptr);
-  int i, j;
+  size_t i, j;
   for (i = 0; i < buflen; i += 16) {
     printf("%06x: ", i);
     for (j = 0; j < 16; j++) {
